pyproject.toml
[project]
name = "wordtolatex-server"
version = "0.1.0"
description = "Word-to-LaTeX conversion API and core library"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.1",
    "lxml>=5.0",
    "Jinja2>=3.1",
    "click>=8.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "pymupdf>=1.23.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.3.0",
    "opencv-python-headless>=4.8.0",
    "python-doctr[torch]>=1.0.0",
    "layoutparser[layoutmodels]>=0.3.4",
    "detectron2@git+https://github.com/facebookresearch/detectron2.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.1",
    "python-docx>=1.1.0",
    "pytest-snapshot>=0.9.0",
    "httpx>=0.27",
]
ml = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "layoutparser>=0.3.0",
    "opencv-python-headless>=4.8.0",
    "ultralytics>=8.0.0",
]
llm = [
    "llama-cpp-python>=0.2.90",
]

[project.scripts]
wordtolatex = "wordtolatex.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/wordtolatex", "src/server"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.uv.extra-build-dependencies]
detectron2 = ["torch", "setuptools", "wheel", "ninja"]
